requirements.txt
autopep8==2.1.0
beautifulsoup4==4.12.3
bs4==0.0.2
fastapi==0.110.0
markovify==0.9.4
pylint==3.1.0
pytest==8.1.1
pytest-datadir==1.5.0
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.31.0
uvicorn==0.28.1
